refactor: extract config into its own crate
