+++ config/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "config"
+version = { workspace = true }
+authors = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+description = { workspace = true }
+repository = { workspace = true }
+
+[features]
+tuning = []
+
+[dependencies]
+uci = { path = "../uci" }
+
+[lints]
+workspace = true
