+++ search/src/lib.rs
@@ -1,7 +1,6 @@
 #![feature(portable_simd)]
 
 mod aspiration;
-mod config;
 mod engine;
 mod history;
 mod lmr;
@@ -19,7 +18,6 @@ mod utils;
 /// Maximum search depth supported by the engine.
 pub const MAX_DEPTH: usize = 100;
 
-pub use config::EngineConfig;
 pub use engine::Engine;
 pub use pv::PvLine;
 pub use result::SearchResult;
