+++ training/Cargo.toml
@@ -8,6 +8,7 @@ description = "NNUE training and data generation tools"
 repository = { workspace = true }
 
 [dependencies]
+config = { path = "../config" }
 search = { path = "../search" }
 nnue = { path = "../nnue" }
 uci = { path = "../uci" }
